Add TeeBind and TeeBindAsync for Result
TeeBind is used to call a function that returns a Result where we want to
ignore the success value of that result (if it succeeds) but, should it
fail, we want to use the failure value.

This is most useful for side-effecting methods that could fail (e.g.
writing to a database).

This implementation will work for any type of success value in the
function, rather than being limited to just unit - whereas in other Tee
implementations we are limited to returning Unit and must use TeeIgnore to
explicitly disregard a different return type. This is done simply because
TeeIgnoreBind is a bit of a mouthful, and TeeIgnoreBindAsync is rather
much for a simple function.